app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
azure-identity
azure-keyvault-secrets
azure-storage-blob
azure-servicebus
opencensus-ext-azure
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
